habr-api/requirements/local.txt
-r base.txt

django-debug-toolbar==1.5
flower==0.9.1
Fabric3==1.12.post1